# Optimize msgp size