* Add messagepack tags.
* Store DataPartInfoNumbers and DataPartInfoSizes as delta-encoded arrays.

```
BEFORE:
pkg: github.com/harshavardhana/xl-meta-bench
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1-32            1428249           812 ns/op       1.23 MB/s        1200 B/op         16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1-32
    xl-meta_test.go:21: msgpack-fast Size: 470 B
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x50-32             46910         30150 ns/op       1.66 MB/s       57848 B/op        702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x50-32
    xl-meta_test.go:21: msgpack-fast Size: 21 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1000-32            4614        267290 ns/op       3.74 MB/s     1152831 B/op      14002 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1000-32
    xl-meta_test.go:21: msgpack-fast Size: 424 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x10000-32            537       4880053 ns/op       2.05 MB/s    11527890 B/op     140004 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x10000-32
    xl-meta_test.go:21: msgpack-fast Size: 4.1 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1-32           1619071           742 ns/op      67.39 MB/s        2016 B/op         16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1-32
    xl-meta_test.go:21: msgpack-fast Size: 768 B
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x50-32            42837         27996 ns/op      89.30 MB/s       98642 B/op        702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x50-32
    xl-meta_test.go:21: msgpack-fast Size: 36 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1000-32           3156        638927 ns/op      78.26 MB/s     1968840 B/op      14002 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1000-32
    xl-meta_test.go:21: msgpack-fast Size: 715 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x10000-32           262       4187965 ns/op     119.39 MB/s    19687905 B/op     140006 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x10000-32
    xl-meta_test.go:21: msgpack-fast Size: 7.0 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1-32          193503          5996 ns/op     166.78 MB/s       17568 B/op         16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1-32
    xl-meta_test.go:21: msgpack-fast Size: 8.0 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x50-32           3998        258938 ns/op     193.10 MB/s      876253 B/op        702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x50-32
    xl-meta_test.go:21: msgpack-fast Size: 399 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1000-32          370       4682133 ns/op     213.58 MB/s    17520958 B/op      14005 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1000-32
    xl-meta_test.go:21: msgpack-fast Size: 7.8 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x10000-32          39      56807636 ns/op     176.03 MB/s    175208761 B/op    140028 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x10000-32
    xl-meta_test.go:21: msgpack-fast Size: 78 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1-32          68949         15856 ns/op     630.68 MB/s      165025 B/op         16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1-32
    xl-meta_test.go:21: msgpack-fast Size: 78 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x50-32          1665        693850 ns/op     720.62 MB/s     8249069 B/op        702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x50-32
    xl-meta_test.go:21: msgpack-fast Size: 3.8 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1000-32          80      25580748 ns/op     390.92 MB/s    164977301 B/op     14015 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1000-32
    xl-meta_test.go:21: msgpack-fast Size: 76 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x10000-32          1    1104248200 ns/op      90.56 MB/s    1649805944 B/op   141052 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x10000-32
    xl-meta_test.go:21: msgpack-fast Size: 765 MiB
PASS

AFTER:

BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1-32         	 2458467	       574 ns/op	   1.74 MB/s	    1200 B/op	      16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1-32
    xl-meta_test.go:23: msgpack-fast Size: 278 B
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x50-32        	   64156	     16043 ns/op	   3.12 MB/s	   57842 B/op	     702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x50-32
    xl-meta_test.go:23: msgpack-fast Size: 13 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1000-32      	    4284	    244453 ns/op	   4.09 MB/s	 1152836 B/op	   14002 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x1000-32
    xl-meta_test.go:23: msgpack-fast Size: 251 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x10000-32     	     498	   3998890 ns/op	   2.50 MB/s	11527892 B/op	  140004 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1x10000-32
    xl-meta_test.go:23: msgpack-fast Size: 2.5 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1-32        	 1625649	       708 ns/op	  70.66 MB/s	    2016 B/op	      16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1-32
    xl-meta_test.go:23: msgpack-fast Size: 380 B
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x50-32       	   37608	     30958 ns/op	  80.76 MB/s	   98642 B/op	     702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x50-32
    xl-meta_test.go:23: msgpack-fast Size: 18 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1000-32     	    2352	    457586 ns/op	 109.27 MB/s	 1968837 B/op	   14002 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x1000-32
    xl-meta_test.go:23: msgpack-fast Size: 351 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x10000-32    	     289	   4578883 ns/op	 109.20 MB/s	19687908 B/op	  140005 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-50x10000-32
    xl-meta_test.go:23: msgpack-fast Size: 3.4 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1-32      	  187458	      5992 ns/op	 166.89 MB/s	   17568 B/op	      16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1-32
    xl-meta_test.go:23: msgpack-fast Size: 2.2 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x50-32     	    6664	    296735 ns/op	 168.50 MB/s	  876255 B/op	     702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x50-32
    xl-meta_test.go:23: msgpack-fast Size: 110 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1000-32   	     288	   3761261 ns/op	 265.87 MB/s	17520964 B/op	   14005 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x1000-32
    xl-meta_test.go:23: msgpack-fast Size: 2.2 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x10000-32  	      44	  59627020 ns/op	 167.71 MB/s	175208705 B/op	  140025 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-1000x10000-32
    xl-meta_test.go:23: msgpack-fast Size: 22 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1-32     	   72710	     15531 ns/op	 643.88 MB/s	  165024 B/op	      16 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1-32
    xl-meta_test.go:23: msgpack-fast Size: 20 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x50-32    	    1737	    676030 ns/op	 739.61 MB/s	 8249065 B/op	     702 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x50-32
    xl-meta_test.go:23: msgpack-fast Size: 989 KiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1000-32  	      98	  25679235 ns/op	 389.42 MB/s	164977216 B/op	   14012 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x1000-32
    xl-meta_test.go:23: msgpack-fast Size: 19 MiB
BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x10000-32 	       5	 204045880 ns/op	 490.09 MB/s	1649775358 B/op	  140211 allocs/op
--- BENCH: BenchmarkParseUnmarshalTinylibMsg/msgpack-fast-10000x10000-32
    xl-meta_test.go:23: msgpack-fast Size: 193 MiB
PASS

```